requirements.txt
streamlit==1.40.1
pandas==1.5.3
streamlit-extras==0.5.0
streamlit-pandas==0.0.9
google-cloud==0.34.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
python==3.11.5
streamlit-option-menu==0.4.0
streamlit==1.40.1
streamlit-extras==0.5.0
plotly.express==0.4.1

